backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
python-dotenv==1.0.0
google-generativeai==0.3.1
flask-jwt-extended==4.5.3
requests==2.31.0
pytest==7.4.3
black==23.11.0
isort==5.12.0
marshmallow==3.20.1
